trying 32-bit floats...

+++ pd_photoz_timed.py
@@ -17,7 +17,7 @@ def mapper1(catalog_dir, nside, ra_col, dec_col, out_dir):
         if cat.endswith(".csv"):
             # read catalog
             try:
-                c = pandas.read_csv(join(catalog_dir, cat), sep=',', low_memory=False, header=1, dtype={ra_col : np.float64, dec_col : np.float64})
+                c = pandas.read_csv(join(catalog_dir, cat), sep=',', low_memory=False, header=1, dtype={ra_col : np.float32, dec_col : np.float32})
                 ra = c.loc[:, ra_col]
                 dec = c.loc[:, dec_col]
             except:
